backend/requirements.txt
Django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-cors-headers
